# server/api/Models/JobApplication.cs
using System.ComponentModel.DataAnnotations;

namespace api.Models;

public enum JobApplicationStatus
{
    Applied,
    Interviewing,
    Rejected,
    Offer,
}

public class JobApplication
{
    public string Id { get; set; } = Guid.NewGuid().ToString();

    [Required]
    public string CompanyName { get; set; } = string.Empty;

    [Required]
    public string Position { get; set; } = string.Empty;

    public JobApplicationStatus Status { get; set; } = JobApplicationStatus.Applied;

    public DateTime DateApplied { get; set; } = DateTime.UtcNow;

    public string UserId { get; set; } = string.Empty;

    public AppUser User { get; set; } = null!;
}